requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
APScheduler==3.10.4
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.1.1
Flask-Login==0.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.31
twilio==8.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
WTForms==3.1.2
yarl==1.9.4
